vendor/omni: Add chromium_prebuilt.sh
This script will pull the necessary libs from a complete build and generate proper
makefiles in order to support using a prebuilt Chromium. Files are placed in
prebuilts/chromium/-devicenamehere-, and are not tracked by git, so they are completely
out of the way.

PS3:
-replace git -C with params compatible to old git versions

Originally commited by: Jiangyi
Adapted for RR by nilac

Change-Id: Iaa13e954054f68f38f8430cdb6185633132249b3